backend/requirements.txt
flask
flask-cors
mysql-connector-python
bcrypt
flask-mail
itsdangerous
python-dotenv
